implementiran detalj da kada god udje u kviz da se nav view oboji u zavisnosti od tacnosti pitanja, popravljen bug za ispis rezultata

+++ app/src/main/java/ba/etf/rma21/projekat/data/repositories/PitanjeKvizRepository.kt
@@ -88,7 +88,7 @@ class PitanjeKvizRepository {
             return pom.getZavrsenKviz()
         }
 
-        fun dodajRezultat(uradjeniKviz: String, uradjeniPredmet: String, rezultat: Int) {
+        fun dodajRezultat(uradjeniKviz: String, uradjeniPredmet: String, rezultat: Double) {
             val pom: PitanjeKviz = listaSvihOdgovorenihPitanja.stream().filter {
                     pitanje ->
                 pitanje.kviz == uradjeniKviz &&
